Output in Daggerfall book format
It is now possible to output a book in the format that the game uses.
This doesn't have much use now, but will be very useful when I add the
ability to read other formats.

Decoding a book and then encoding it will not produce a completely
identical file, but it should look exactly the same.